+++ samples/Curiosity.Tools.Sample/Program.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Threading.Tasks;
+using Curiosity.Tools.Performance;
 
 namespace Curiosity.Tools.Sample
 {
@@ -14,7 +16,17 @@ namespace Curiosity.Tools.Sample
             
             Console.WriteLine($"\nResult of {nameof(PhoneHelperSample)}:");
             PhoneHelperSample.Run();
-            
+
+
+            using (PerformanceManager.Measure("Test"))
+            {
+                Task.Delay(TimeSpan.FromSeconds(1));
+            }
+
+            using (StuckCodeManager.Enter("Stuck code", 1))
+            {
+                Task.Delay(TimeSpan.FromSeconds(2));
+            }
         }
     }
 }