+++ src/Curiosity.Tools/Performance/PerformanceMeasurer.cs
@@ -1,0 +1,45 @@
+using System;
+using System.Diagnostics;
+using Microsoft.Extensions.Logging;
+
+namespace Curiosity.Tools.Performance
+{
+    /// <summary>
+    /// Class for measuring performance.
+    /// </summary>
+    public class PerformanceMeasurer : IDisposable
+    {
+        private readonly ILogger? _logger;
+        private readonly Stopwatch _stopwatch;
+        private string _name;
+
+        internal PerformanceMeasurer(ILogger? logger, string name)
+        {
+            _logger = logger;
+            _name = name;
+            _stopwatch = Stopwatch.StartNew();
+        }
+
+        /// <inheritdoc />
+        public void Dispose()
+        {
+            if (_stopwatch.IsRunning)
+            {
+                _stopwatch.Stop();
+                _logger?.LogInformation("{0} -> {1} ms.", _name, _stopwatch.ElapsedMilliseconds);
+            }
+        }
+
+        /// <summary>
+        /// Replaces name of measurer for a new one.
+        /// </summary>
+        /// <param name="name">Name of format for name.</param>
+        /// <param name="args">Arguments for name format.</param>
+        public void ReplaceName(string name, params object[]? args)
+        {
+            _name = args != null 
+                ? String.Format(name, args)
+                : name;
+        }
+    }
+}