+++ src/Curiosity.Hosting/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.4] - 2021-03-15
+
+### Added 
+
+- Added IoC extensions and initializer of measurer.
+- Added performance measurer and stuck code manager to bootstrappers.
+
 ## [1.1.3] - 2021-03-05
 
 ### Added